48 started reading image display refactor
