+++ Assets/Scripts/InteractableObjects/ReadingImageDisplay.cs.meta
